Move character classes to top level
